Rewrite line wrapping to use a linear-time algo.
Encoded bytes are moved from the end to the front so each byte is
only moved once.

Encoding is somewhat rearranged to operate on a slice into the
output buffer. This makes it easier to avoid clobbering any
existing bytes in the buffer, as well as paving the way to slice-
based encoding needed for a Display wrapper, stream adapters, etc.