Don't allocate memory within decode_config.
decode_config invokes decode_config_buf which resizes the vector to the
correct size. There is no reason for decode_config to attempt to
pre-allocate the buffer. In the best case it moves the allocation from
within decode_config_buf to within decode_config, but unfortunately
there is a bug in the calculation used by decode_config that means it
underallocates for tiny buffers (requiring decode_config_buf to allocate
again), and allocates way more memory than requires for larger buffers.