Trait based configuration.
The configuration consists of 3 traits.

Encoding: Determine how to encode a 6-bit value into an 8-bit value.
Decoding: Determine how to decode an 8-bit value into a 6-bit value.
Padding: Whether to use padding and if so which character to use.

These traits are currently sealed so they cannot be implemented outside
this crate. This could be changed in the future, but until the trait
definitions get a little more flushed out it seemed wise to not have
users relying on the current definition.

Each of the 3 supported alphabets defines it's own zero sized type and
implements the traits.

Public API methods are now converted to be generically bounded by the
traits which will monomorphize each function specifically for the
desired configuration. This results in a small performance improvement.

This method also provides a path to supporting custom alphabets in a
reasonable way. That functionality will be done in a future commit.