Provide the ability to encode and decode with a custom config.
There is now a CustomConfig type. It supports custom alphabets and
padding and can be built with a builder pattern.

// A custom config that matches the builtin Crypt alphabet.
let custom_config = CustomConfigBuilder::with_alphabet(
    b"./0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz"
)
.with_padding(b'=')
.build()?;

The alphabet and padding provided must be ascii characters. This
limitation is in place to ensure that encoded values are valid UTF-8 and
can therefore be represented as rust String values.

An immutable pointer to CustomConfig implements Encoding + Decoding +
Padding. This allows it to be used in contexts where the standard
builtin configs are used.

Creating a CustomConfig is somewhat heavyweight and the resulting type
is somewhat large (~2.5KB). It is expected that building the config
would be early in program execution and that an immutable reference is
what will be efficiently passed around. i.e. The configuration would be a
good candidate to be used with lazy_static.