Make EncoderWriter own the underlying writer.
Previously EncoderWriter stored a mutable reference to the generic
writer, which seems intuitive but I believe is an anti-pattern. Instead,
it's more flexible to take ownership of the underlying writer.

The standard library has a blanket impl<W> Write for &mut W. This allows
users that want to retain ownership of the underlying writer to pass a
mutable reference into the EncoderWriter, but users that want to
relinquish ownership can pass the underlying writer by value and it will
be dropped when EncoderWriter is dropped.