Advertise that unsafe code is forbidden
Having zero unsafe code and forbidding it is a desirable property for Rust code. This commit adds a badge to advertise this property. This helps users make a more informed choice when picking a base64 crate.

In this PR clicking the badge leads to the Rust Secure Code WG coordination repo for removing unnecessary unsafe code from the Rust ecosystem. I'm fine with making the link lead somewhere else or making the badge non-clickable if you prefer.