Roll back the MSRV
The current MSRV (for non-dev builds) is Rust 1.57 however the codebase
does not use any features from edition 2021 (landed Rust 1.56) except
the `TryFrom` trait.

If we explicitly import `core::convert::tryFrom` we can stay on edition
2018.

The `rust-version` manifest option was added in Rust 1.56 but lets leave
it there as added documentation of the MSRV and in preparation for
bumping the MSRV when needed.

Rust 1.48 is a suitably old toolchain version, it is in Debian
oldstable.

Roll back the crate MSRV to Rust 1.48