`no_std` import simplification.
It's always OK to use `core` and `alloc` if you're using `std`, and
`core` if you're using `alloc`. Therefore, code can be simplified by
using `alloc` even if `std` is also being used, and so on.

* Make `std` feature depend on `alloc` feature.
* Simplify `cfg()`s by making that assumption.
* Don't import `std` aliased as `alloc`; use `alloc` directly.

These changes do not affect the public API of the crate in any way.